chore: add rpc hook
